+++ app/src/main/kotlin/com/wire/android/ui/home/conversations/ConversationScreen.kt
@@ -79,11 +79,10 @@ import com.wire.android.ui.home.conversations.messages.ConversationMessagesViewS
 import com.wire.android.ui.home.conversations.model.AssetBundle
 import com.wire.android.ui.home.conversations.model.EditMessageBundle
 import com.wire.android.ui.home.conversations.model.UIMessage
-import com.wire.android.ui.home.conversations.selfdeletion.SelfDeletionMenuItems
 import com.wire.android.ui.home.messagecomposer.MessageComposer
-import com.wire.android.ui.home.messagecomposer.state.MessageComposerState
+import com.wire.android.ui.home.messagecomposer.state.MessageComposerInnerState
 import com.wire.android.ui.home.messagecomposer.model.UiMention
-import com.wire.android.ui.home.messagecomposer.state.rememberMessageComposerState
+import com.wire.android.ui.home.messagecomposer.state.rememberMessageComposerInnerState
 import com.wire.android.ui.home.newconversation.model.Contact
 import com.wire.android.util.permission.CallingAudioRequestFlow
 import com.wire.android.util.permission.rememberCallingRecordAudioBluetoothRequestFlow
@@ -313,7 +312,7 @@ private fun ConversationScreen(
     tempWritableVideoUri: Uri?
 ) {
     val conversationScreenState = rememberConversationScreenState()
-    val messageComposerInnerState = rememberMessageComposerState()
+    val messageComposerInnerState = rememberMessageComposerInnerState()
     val context = LocalContext.current
 
     LaunchedEffect(conversationMessagesViewModel.savedStateHandle) {
@@ -359,10 +358,11 @@ private fun ConversationScreen(
             }
 
             is ConversationScreenState.BottomSheetMenuType.SelfDeletion -> {
-                SelfDeletionMenuItems(
-                    currentlySelected = messageComposerInnerState.selfDeletionDuration,
-                    onSelfDeletionDurationChanged = { messageComposerInnerState.specifySelfDeletionTime(it) }
-                )
+                emptyList()
+//                SelfDeletionMenuItems(
+//                    currentlySelected =
+//                    onSelfDeletionDurationChanged = { messageComposerInnerState.specifySelfDeletionTime(it) }
+//                )
             }
 
             ConversationScreenState.BottomSheetMenuType.None -> emptyList()
@@ -444,7 +444,7 @@ private fun ConversationScreenContent(
     lastUnreadMessageInstant: Instant?,
     conversationState: ConversationViewState,
     audioMessagesState: Map<String, AudioState>,
-    messageComposerState: MessageComposerState,
+    messageComposerState: MessageComposerInnerState,
     messages: Flow<PagingData<UIMessage>>,
     onSendMessage: (String, List<UiMention>, String?, expireAfter : Duration?) -> Unit,
     onSendEditMessage: (EditMessageBundle) -> Unit,
